Fixing bugs and implement features
We need to revise the resposive css structure - the #mobile with display
none is used for html element ids - It is not good to have repeated ids.
Check how .notebook is used - is a css class, not an id.

Check made slide n.5 changes in html - propably it will be a model to
other slides with videos.

fix slide resize when resizing window

fix video playback issues regarding play/pause

Added mobile touch support - tested only in browser simulator